# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management System API
  description: |
    API для управления банковскими картами, пользователями и транзакциями.
    Включает аутентификацию, работу с картами, переводы средств и административные функции.
  version: 1.0.0
  contact:
    name: API Support
    email: support@bankcards.com
  license:
    name: Proprietary
    url: https://www.bankcards.com/terms
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      operationId: authenticateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: User authenticated successfully
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  type: Bearer
                  id: 1
                  username: testuser
                  email: user@example.com
                  role: USER
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация
      description: Регистрация нового пользователя
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Некорректные данные регистрации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Текущий пользователь
      description: Получить информацию о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешно получена информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Current user retrieved
                data:
                  id: 1
                  username: testuser
                  firstName: Ivan
                  lastName: Ivanov
                  email: ivan@example.com
                  role: USER
                  enabled: true
                  createdAt: '2024-01-15T10:30:00'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards:
    get:
      tags:
        - Cards
      summary: Получить карты пользователя
      description: Получить список карт текущего пользователя с пагинацией и фильтрацией
      operationId: getUserCards
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Статус карты для фильтрации
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED, PENDING]
        - name: cardNumberLastFour
          in: query
          description: Последние 4 цифры номера карты
          schema:
            type: string
            pattern: '^\d{4}$'
        - name: ownerName
          in: query
          description: Имя владельца карты
          schema:
            type: string
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Поле для сортировки
          schema:
            type: string
            default: createdAt
        - name: sortDirection
          in: query
          description: Направление сортировки
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCardResponse'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/generate:
    post:
      tags:
        - Cards
      summary: Сгенерировать новую карту
      description: Создать новую карту для текущего пользователя
      operationId: generateCard
      security:
        - bearerAuth: []
      parameters:
        - name: ownerName
          in: query
          required: true
          description: Имя владельца карты
          schema:
            type: string
            example: IVAN IVANOV
      responses:
        '201':
          description: Карта успешно сгенерирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/{cardId}:
    get:
      tags:
        - Cards
      summary: Получить карту по ID
      description: Получить информацию о конкретной карте
      operationId: getCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          description: ID карты
          schema:
            type: integer
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/{cardId}/balance:
    get:
      tags:
        - Cards
      summary: Получить баланс карты
      description: Получить текущий баланс карты
      operationId: getCardBalance
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Баланс карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Balance retrieved successfully
                data: 1500.75
  /api/cards/{cardId}/activate:
    post:
      tags:
        - Cards
      summary: Активировать карту
      description: Активировать заблокированную карту (если не истек срок)
      operationId: activateCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/cards/{cardId}/request-block:
    post:
      tags:
        - Cards
      summary: Запросить блокировку карты
      description: Пользователь запрашивает блокировку своей карты
      operationId: requestBlockCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
        - name: reason
          in: query
          description: Причина блокировки
          schema:
            type: string
      responses:
        '200':
          description: Запрос на блокировку успешно отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Карта не активна или другие ошибки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/transactions/card/{cardId}:
    get:
      tags:
        - Transactions
      summary: Получить транзакции карты
      description: Получить историю транзакций конкретной карты
      operationId: getCardTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Транзакции карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionResponse'
...
  /api/admin/cards/{cardId}:
    get:
      tags:
        - Admin Cards
      summary: Получить карту по ID
      description: Получить информацию о любой карте в системе - только для администратора
      operationId: adminGetCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о карте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags:
        - Admin Cards
      summary: Удалить карту
      description: Удалить карту из системы (только если баланс = 0) - только для администратора
      operationId: deleteCard
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Нельзя удалить карту с ненулевым балансом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: testuser
        password:
          type: string
          example: password
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - firstName
        - lastName
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: danila
        password:
          type: string
          minLength: 6
          example: Danila12345
        firstName:
          type: string
          example: Danila
        lastName:
          type: string
          example: Testov
        email:
          type: string
          format: email
          example: danila@example.com
...
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        maskedNumber:
          type: string
          example: '**** **** **** 9843'
        ownerName:
          type: string
          example: IVAN IVANOV
        expiryDate:
          type: string
          example: '12/27'
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED, PENDING]
          example: ACTIVE
        balance:
          type: number
          format: decimal
          example: 1500.75
        expired:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00'
        userId:
          type: integer
          example: 1
        username:
          type: string
          example: testuser
...
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: 'txn_7s8df9s7d8f9s'
        amount:
          type: number
          format: decimal
          example: 500.50
        currency:
          type: string
          example: RUB
        type:
          type: string
          example: TRANSFER
        status:
          type: string
          example: COMPLETED
        description:
          type: string
          example: Перевод между картами
        fromCardMasked:
          type: string
          example: '**** **** **** 8597'
        toCardMasked:
          type: string
          example: '**** **** **** 0065'
        transactionDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Response обертки
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Данные ответа
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: Detailed error description
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00'
        path:
          type: string
          example: '/api/cards'
    PaginatedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
    PaginatedCardResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/CardResponse'
    PaginatedTransactionResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TransactionResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен аутентификации
security:
  - bearerAuth: []